# web/content/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProxyWhirl API",
    "description": "REST API for advanced proxy rotation with auto-fetching, validation, and persistence. Manage proxy pools, make proxied requests, and monitor system health through RESTful endpoints.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/config": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "Get current configuration",
        "description": "Get current API configuration.\n\nArgs:\n    config: Configuration dependency\n    api_key: API key verification\n\nReturns:\n    Current configuration settings",
        "operationId": "get_configuration_api_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ConfigurationSettings_"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      },
      "put": {
        "tags": [
          "Configuration"
        ],
        "summary": "Update configuration",
        "description": "Update API configuration at runtime.\n\nArgs:\n    update_data: Configuration updates (partial)\n    config: Configuration dependency\n    rotator: ProxyWhirl dependency\n    api_key: API key verification\n\nReturns:\n    Updated configuration settings",
        "operationId": "update_configuration_api_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ConfigurationSettings_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/retry/policy": {
      "get": {
        "tags": [
          "Retry & Failover"
        ],
        "summary": "Get global retry policy",
        "description": "Get the current global retry policy configuration.\n\nReturns:\n    Current retry policy settings",
        "operationId": "get_retry_policy_api_retry_policy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_RetryPolicyResponse_"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      },
      "put": {
        "tags": [
          "Retry & Failover"
        ],
        "summary": "Update global retry policy",
        "description": "Update the global retry policy configuration.\n\nArgs:\n    policy_request: New retry policy settings\n\nReturns:\n    Updated retry policy",
        "operationId": "update_retry_policy_api_retry_policy_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_RetryPolicyResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/circuit-breakers": {
      "get": {
        "tags": [
          "Retry & Failover"
        ],
        "summary": "List all circuit breaker states",
        "description": "Get circuit breaker states for all proxies.\n\nReturns:\n    List of circuit breaker states",
        "operationId": "list_circuit_breakers_api_circuit_breakers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_CircuitBreakerResponse__"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/api/circuit-breakers/{proxy_id}": {
      "get": {
        "tags": [
          "Retry & Failover"
        ],
        "summary": "Get circuit breaker state for specific proxy",
        "description": "Get circuit breaker state for a specific proxy.\n\nArgs:\n    proxy_id: Proxy ID to get circuit breaker for\n\nReturns:\n    Circuit breaker state",
        "operationId": "get_circuit_breaker_api_circuit_breakers__proxy_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "proxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proxy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_CircuitBreakerResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/circuit-breakers/{proxy_id}/reset": {
      "post": {
        "tags": [
          "Retry & Failover"
        ],
        "summary": "Manually reset circuit breaker",
        "description": "Manually reset a circuit breaker to CLOSED state.\n\nArgs:\n    proxy_id: Proxy ID whose circuit breaker to reset\n\nReturns:\n    Updated circuit breaker state",
        "operationId": "reset_circuit_breaker_api_circuit_breakers__proxy_id__reset_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "proxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proxy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_CircuitBreakerResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_ConfigurationSettings_": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Response status: 'success' or 'error'"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfigurationSettings"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload (present on success)"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error details (present on failure)"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo",
            "description": "Response metadata (request ID, timestamp, version)"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "APIResponse[ConfigurationSettings]",
        "examples": [
          {
            "data": {
              "message": "Operation completed successfully"
            },
            "meta": {
              "request_id": "550e8400-e29b-41d4-a716-446655440000",
              "timestamp": "2025-10-27T12:00:00Z",
              "version": "1.0.0"
            },
            "status": "success"
          },
          {
            "error": {
              "code": "VALIDATION_ERROR",
              "details": {
                "field": "url",
                "reason": "Invalid URL format"
              },
              "message": "Invalid request parameters",
              "timestamp": "2025-10-27T12:00:00Z"
            },
            "meta": {
              "request_id": "550e8400-e29b-41d4-a716-446655440001",
              "timestamp": "2025-10-27T12:00:00Z",
              "version": "1.0.0"
            },
            "status": "error"
          }
        ]
      },
...
      "APIResponse_RetryPolicyResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Response status: 'success' or 'error'"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetryPolicyResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload (present on success)"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error details (present on failure)"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo",
            "description": "Response metadata (request ID, timestamp, version)"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "APIResponse[RetryPolicyResponse]",
        "examples": [
          {
            "data": {
              "message": "Operation completed successfully"
            },
            "meta": {
              "request_id": "550e8400-e29b-41d4-a716-446655440000",
              "timestamp": "2025-10-27T12:00:00Z",
              "version": "1.0.0"
            },
            "status": "success"
          },
          {
            "error": {
              "code": "VALIDATION_ERROR",
              "details": {
                "field": "url",
                "reason": "Invalid URL format"
              },
              "message": "Invalid request parameters",
              "timestamp": "2025-10-27T12:00:00Z"
            },
            "meta": {
              "request_id": "550e8400-e29b-41d4-a716-446655440001",
              "timestamp": "2025-10-27T12:00:00Z",
              "version": "1.0.0"
            },
            "status": "error"
          }
        ]
      },
...
      "APIResponse_list_CircuitBreakerResponse__": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Response status: 'success' or 'error'"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CircuitBreakerResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload (present on success)"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error details (present on failure)"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo",
            "description": "Response metadata (request ID, timestamp, version)"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "APIResponse[list[CircuitBreakerResponse]]",
        "examples": [
          {
            "data": {
              "message": "Operation completed successfully"
            },
            "meta": {
              "request_id": "550e8400-e29b-41d4-a716-446655440000",
              "timestamp": "2025-10-27T12:00:00Z",
              "version": "1.0.0"
            },
            "status": "success"
          },
          {
            "error": {
              "code": "VALIDATION_ERROR",
              "details": {
                "field": "url",
                "reason": "Invalid URL format"
              },
              "message": "Invalid request parameters",
              "timestamp": "2025-10-27T12:00:00Z"
            },
            "meta": {
              "request_id": "550e8400-e29b-41d4-a716-446655440001",
              "timestamp": "2025-10-27T12:00:00Z",
              "version": "1.0.0"
            },
            "status": "error"
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RateLimitConfig": {
        "properties": {
          "default_limit": {
            "type": "integer",
            "title": "Default Limit",
            "default": 100
          },
          "request_endpoint_limit": {
            "type": "integer",
            "title": "Request Endpoint Limit",
            "default": 50
          }
        },
        "type": "object",
        "title": "RateLimitConfig",
        "description": "Rate limiting configuration."
      },
...
      "UpdateConfigRequest": {
        "properties": {
          "rotation_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rotation Strategy"
          },
          "timeout": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 300,
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout"
          },
          "max_retries": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10,
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Retries"
          },
          "rate_limits": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RateLimitConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "cors_origins": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cors Origins"
          }
        },
        "type": "object",
        "title": "UpdateConfigRequest",
        "description": "Request to update API configuration (partial updates allowed).",
        "example": {
          "rotation_strategy": "round-robin",
          "timeout": 60
        }
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}